♻️ ai: split shared types (Message/StreamEvent/Usage/...) out of provider.go into types.go
